Add files via upload
Changes made in this commit include converting the logic of all pieces so that their next possible moves are determined based on the game object rather than on DOM elements. Also, I have started a checkmate logic, but so far I have only being able to tweak the possible moves a piece has when the king is on check. Notable future additions would include a logic to determine that moving a piece of a particular color wouldn't be putting that piece's own king on check, and finally discovering checkmate. I still haven't added the en passant logic to the pawn's.